# Conflicts:
#	server/src/main/java/de/pnp/manager/component/crafting/CraftingRecipe.java
#	server/src/test/java/de/pnp/manager/server/database/CraftingRecipeRepositoryTest.java

+++ server/src/main/java/de/pnp/manager/component/item/interfaces/IDefensiveItem.java
@@ -5,7 +5,7 @@ package de.pnp.manager.component.item.interfaces;
  */
 public interface IDefensiveItem extends IDamageableItem, IEquipableItem {
 
-    double getWeight();
+    float getWeight();
 
     int getArmor();
 }
